requirements.txt
kafka-python
pyspark[connect]==3.5.0